perf: lazy-load Sentry SDK via dynamic import (admin-only)
Sentry (with Replay) is ~174KB raw / 53KB gzip in common.min.js (35% of
the bundle), statically imported from o11y/instrument.ts, error.tsx and
ErrorBoundary.tsx even though it is only ever used when isAdmin() is
true. Replacing the static imports with a lazy wrapper (react/o11y/sentry.ts)
that dynamic-imports '@sentry/react' lets webpack split it into its own
chunk, so storefront (non-admin) requests never fetch or parse it.

EXPERIMENTAL - testing on storecomponents/qareview before deciding whether
to pursue upstream.
